Change: Update python workflows
Extract codecov workflow. Codecov is unreliable and should not be run in
a matrix. Therefore allow to include and run it separately.

Drop matrix from ci-python workflow. This allows to specify the matrix
in the calling workflow.

Add mypy to ci-python workflow. Allow to run mypy via the ci-python
workflow.