Change: Update cloud release workflow
Cloud will always create calendar based releases and tags without a
leading 'v'. Additionally the cloud workflow should not update the
version after a release to not create an additional commit.
All releases for the cloud will be marked as pre-releases initially and
are changed to "real" releases manually after testing the release.