Add: Allow to override build-args for container build 2nd gen
The calling workflow should be able to specify the build-args. For
example it should be possible to use whatever gvm-libs base image
version the calling workflow needs and shouldn't limit to oldstable and
oldstable-edge.

The change tries to be backwards compatible to still provide the old
values if no build-args are passed.